requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
optuna>=3.3.0
PyYAML>=6.0
matplotlib>=3.7.0
